+++ irc/client.py
@@ -64,7 +64,6 @@ import itertools
 import contextlib
 
 import six
-from tempora import schedule
 from jaraco.itertools import always_iterable, infinite_call
 from jaraco.functools import Throttler
 from jaraco.stream import buffer
@@ -81,6 +80,7 @@ from . import functools as irc_functools
 from . import features
 from . import ctcp
 from . import message
+from . import schedule
 
 log = logging.getLogger(__name__)
 
@@ -107,36 +107,6 @@ class PrioritizedHandler(
     def __lt__(self, other):
         "when sorting prioritized handlers, only use the priority"
         return self.priority < other.priority
-
-
-@six.add_metaclass(abc.ABCMeta)
-class IScheduler(object):
-    @abc.abstractmethod
-    def execute_every(self, period, func):
-        "execute func every period"
-
-    @abc.abstractmethod
-    def execute_at(self, when, func):
-        "execute func at when"
-
-    @abc.abstractmethod
-    def execute_after(self, delay, func):
-        "execute func after delay"
-
-    @abc.abstractmethod
-    def run_pending(self):
-        "invoke the functions that are due"
-
-
-class DefaultScheduler(schedule.InvokeScheduler, IScheduler):
-    def execute_every(self, period, func):
-        self.add(schedule.PeriodicCommand.after(period, func))
-
-    def execute_at(self, when, func):
-        self.add(schedule.DelayedCommand.at(when, func))
-
-    def execute_after(self, delay, func):
-        self.add(schedule.DelayedCommand.after(delay, func))
 
 
 class Reactor(object):
@@ -185,7 +155,7 @@ class Reactor(object):
     are guarded by a mutex.
     """
 
-    scheduler_class = DefaultScheduler
+    scheduler_class = schedule.DefaultScheduler
 
     def __do_nothing(*args, **kwargs):
         pass
@@ -211,7 +181,7 @@ class Reactor(object):
         self._on_disconnect = on_disconnect
 
         scheduler = self.scheduler_class()
-        assert isinstance(scheduler, IScheduler)
+        assert isinstance(scheduler, schedule.IScheduler)
         self.scheduler = scheduler
 
         self.connections = []